add checkout page, fix database, save data OrderWeb, OrderWebDetail

+++ src/main/java/com/sonht/e_commerce_webapp_spring_boot/service/CartService.java
@@ -13,5 +13,5 @@ public interface CartService {
 
     void removeAllCartItems();
 
-    Double calculateTotalPrice(List<CartItem> cartItems);
+    Long calculateTotalPrice(List<CartItem> cartItems);
 }
